+++ src/components/LetterForm.jsx
@@ -2,6 +2,10 @@ import { memberKoreanMap } from "lib/member";
 import styled from "styled-components";
 import Button from "./Button";
 
+const NICKNAME = "nickname";
+const CONTENT = "content";
+const SELECTED = "selected";
+
 export default function LetterForm({
   members,
   formState,
@@ -9,8 +13,6 @@ export default function LetterForm({
   handleEnrollLetter,
 }) {
   const { nickname, content } = formState;
-  const handleEtargetValue = (key) => (e) =>
-    handleChangeFormValue(key)(e.target.value);
   return (
     <StyledForm onSubmit={handleEnrollLetter}>
       <StyledInputWrapper>
@@ -21,7 +23,7 @@ export default function LetterForm({
             name='content'
             required
             value={nickname}
-            onChange={handleEtargetValue("nickname")}
+            onChange={handleChangeFormValue(NICKNAME)}
             placeholder='최대 20글자까지만 작성할 수 있습니다'
           />
         </StyledRow>
@@ -34,24 +36,27 @@ export default function LetterForm({
             rows='5'
             required
             value={content}
-            onChange={handleEtargetValue("content")}
+            onChange={handleChangeFormValue(CONTENT)}
             maxLength={100}
             placeholder='최대 100자까지만 작성할 수 있습니다.'></textarea>
         </StyledRow>
       </StyledInputWrapper>
-      <div className='select-wrapper'>
+      <StyledSelectWrapper>
         <label htmlFor='select-memeber'>누구에게 보내실 건가요?</label>
-        <select onChange={handleEtargetValue("selected")}>
-          {members.map((member) => (
-            <option key={"option/" + member} value={member}>
-              {memberKoreanMap[member]}
-            </option>
-          ))}
+        <select id='select-member' onChange={handleChangeFormValue(SELECTED)}>
+          <SelectOptions members={members} />
         </select>
-      </div>
+      </StyledSelectWrapper>
       <Button type='submit'>팬레터 등록</Button>
     </StyledForm>
   );
+}
+function SelectOptions({ members }) {
+  return members.map((member) => (
+    <option key={"options-" + member} value={member}>
+      {memberKoreanMap[member]}
+    </option>
+  ));
 }
 
 const StyledForm = styled.form`
@@ -68,12 +73,12 @@ const StyledForm = styled.form`
   button {
     align-self: flex-end;
   }
-
-  .select-wrapper label {
+  border-radius: 0.5rem;
+`;
+const StyledSelectWrapper = styled.div`
+  label {
     margin-right: 1rem;
   }
-
-  border-radius: 0.5rem;
 `;
 const StyledInputWrapper = styled.div`
   display: flex;
